pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ramp-motion"
version = "0.1.0"
description = "Motion-gated ref/query frame capture for ramp monitoring (DeepStream 6.3)"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0,<3.0",
    "PyYAML>=6.0",
    "numpy>=1.24,<2.0",
    "opencv-python-headless>=4.8,<5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
include = ["ramp_motion*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: test requires CUDA GPU and compiled .so",
    "integration: test requires running DeepStream pipeline",
]
